\name{plr-internal}
\title{
  Internal plr functions
}
\alias{anova.stepplr}
\alias{cross.imat}
\alias{get.imat}
\alias{imat}
\alias{print.plr}
\alias{print.stepplr}
\alias{summary.plr}
\alias{summary.stepplr}
\alias{term.match}
\usage{
  anova.stepplr(object, ...)
  cross.imat(imat1, imat2)
  get.imat(x, level = NULL)
  imat(object, x)
  print.plr(x, ...)
  print.stepplr(x, ...)
  summary.plr(object, ...)
  summary.stepplr(object, ...)
  term.match(term1, term2, termlist, len)
}
\description{
  Internal plr functions
}
\references{
  Mee Young Park and Trevor Hastie (2008) Penalized Logistic Regression
  for Detecting Gene Interactions
}
\author{Mee Young Park and Trevor Hastie}
\keyword{internal}
