\name{plr-internal}
\title{
  Internal plr functions
}
\alias{anova.stepplr}
\alias{cross.Imat}
\alias{get.Imat}
\alias{Imat}
\alias{print.plr}
\alias{print.stepplr}
\alias{summary.plr}
\alias{summary.stepplr}
\alias{term.match}
\usage{
  anova.stepplr(object, ...)
  cross.Imat(imat1, imat2)
  get.Imat(x, level = NULL)
  Imat(object, x)
  print.plr(x, ...)
  print.stepplr(x, ...)
  summary.plr(object, ...)
  summary.stepplr(object, ...)
  term.match(term1, term2, termlist, len)
}
\description{
  Internal plr functions
}
\references{
  Mee Young Park and Trevor Hastie (2006) Penalized Logistic Regression
  for Detecting Gene Interactions - available at the authors' websites,
  \url{http://www.stanford.edu/~mypark} or
  \url{http://stat.stanford.edu/~hastie/pub.htm}.
}
\author{Mee Young Park and Trevor Hastie}
\keyword{internal}
